requirements.txt
Django==1.7.1
MySQL-python==1.2.5
Pillow==2.6.0
South==1.0
argparse==1.2.1
coverage==3.7.1
django-admin-bootstrapped==2.0.5
django-bootstrap3==4.11.0
django-registration-redux==1.1
stevedore==0.15
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
wsgiref==0.1.2
yolk==0.4.3
